Add comment about useOverlayOffset Add offset constant into popover Move overlay types into another types.tsx to prevent circular dependency warning
