+++ src/types.tsx
@@ -30,3 +30,10 @@ export type Color =
   | 'light'
   | 'white'
   | 'muted';
+
+export type Placement = import('react-overlays/usePopper').Placement;
+
+export type ArrowProps = {
+  ref: React.RefCallback<HTMLElement>;
+  style: React.CSSProperties;
+};
