+++ src/useOverlayOffset.tsx
@@ -2,7 +2,10 @@ import { useMemo, useRef } from 'react';
 import hasClass from 'dom-helpers/hasClass';
 import { Options } from 'react-overlays/usePopper';
 import { useBootstrapPrefix } from './ThemeProvider';
+import Popover from './Popover';
 
+// This is meant for internal use.
+// This applies a custom offset to the overlay if it's a popover.
 export default function useOverlayOffset(): [
   React.RefObject<HTMLElement>,
   Options['modifiers'],
@@ -19,9 +22,7 @@ export default function useOverlayOffset(): [
             overlayRef.current &&
             hasClass(overlayRef.current, popoverClass)
           ) {
-            // Default popover offset.
-            // https://github.com/twbs/bootstrap/blob/5c32767e0e0dbac2d934bcdee03719a65d3f1187/js/src/popover.js#L28
-            return [0, 8];
+            return Popover.POPPER_OFFSET;
           }
           return [0, 0];
         },
